refactor(skill): split skill_castend_damage_id and skill_castend_pos2
Move the damage and ground castend switches out of skill.c (Issue #56 phase 2). No behavior change.
